+++ spotdl/download/downloadManager.py
@@ -13,193 +13,21 @@ Everything that has to do with multi-processing is in this file.
 import multiprocess
 from multiprocess import Pool, Lock
 from multiprocess.managers import BaseManager
-import time
-from urllib.request import urlopen
 
 #! The following are not used, they are just here for static typechecking with mypy
 from typing import List
 
+from spotdl.download.downloader import download_song
 from spotdl.search.songObj import SongObj
 from spotdl.download.progressHandlers import DownloadTracker
-from spotdl.download.downloader import download_song
+from spotdl.download.messageHandlers import ParentMessageTracker
+
 
 
 from queue import Queue
 import sys
-from collections import defaultdict
-from datetime import datetime
 
 
-#============================================
-#=== Subprocess Reporting to Main Process ===
-#============================================
-
-class BasicDisplayMessenger():
-    '''
-    This Class is an placeholder for the shared message_queue & overall progress information.
-    It is also responsible for dispatching Process Trackers.
-    '''
-
-    def __init__(self, message_queue, lock = None):
-        self.message_queue = message_queue
-        self.overallProgress = 0
-        self.overallTotal = 100
-        self.lock = lock
-
-    def set_song_count_to(self, count: int):
-        '''
-        `count` : `int` The numbers of songs the current download Pool has mapped
-
-        Sets the current song count so that the DisplayManager can estimate an overall progress indicator.
-        '''
-        self.overallTotal = 100 * count
-        self.put(0, 'Song Count', count)
-
-    def put(self, ID: int, name: str, progress, message = ''):
-        '''
-        `ID` : `int` Current process ID
-
-        `name` : `str` Display Name
-
-        `progress` : 'int` 0-100 
-
-        `message` : `str` (Optional) Message to display
-        
-        queue -> `{ (processID): { 'time': (timestamp), 'name': (display name), 'progress': (0-100), 'message': (message) } }`
-        '''
-
-        message = { ID: { 'time': datetime.timestamp(datetime.now()), 'name': name, 'progress': int(progress), 'message': message}}
-        self.message_queue.put(message)
-
-    def newMessageTracker(self, name):
-        '''
-        Dispatch a new Message Plugin
-
-        `RETURNS` `DownloadMessagesPlugin` instance with `self` (`put()` method) already passed into it.
-        '''
-        return DownloadMessagesPlugin(self, name)
-    
-class DownloadMessagesPlugin():
-    '''
-    A Plugin for handling callback events from each download process seperately. 
-    Each function here corresponds to a function that is called throughout the main download process.
-    This class handles things such as displayName, processProgress, processMessages, etc.
-    A New message Tracker object is created for each download process and passed into it upon creation.
-    '''
-
-    def __init__(self, parent, displayName):
-        self.parent = parent
-        self.displayName = displayName
-        self.ID = None
-        self.progress = 0
-        self.isComplete = False
-        # print('New Tracker Initialized')
-        # self.total = 100
-
-    def notify_download_begin(self, ID: int):
-        '''
-        `ID` : `int` Process ID of download process
-
-        RETURNS `~`
-
-        Should be called at the beginning of each download
-        '''
-        self.ID = ID
-        self.send_update('Download Started')
-
-
-    def set_song_count_to(self, songCount: int) -> None:
-        '''
-        `songCount` : `int` number of songs being downloaded
-
-        RETURNS `~`
-
-        sets the size of the progressbar based on the number of songs in the current
-        download set
-        '''
-
-        #! all calculations are based of the arbitrary choice that 1 song consists of
-        #! 100 steps/points/iterations 
-
-        self.parent.overallTotal = songCount * 100
-        self.send_update('Song Count now' + str(songCount))
-
-    def notify_download_skip(self) -> None:
-        '''
-        updates progress bar to reflect a song being skipped
-        '''
-
-        # self.progressBar.set_by_percent(100)
-        self.progress = 100
-        self.isComplete = True
-        self.send_update('Skipping')
-
-    def download_progress_hook(self, stream, chunk, bytes_remaining) -> None:
-        '''
-        Progress hook built according to pytube's documentation. It is called each time
-        bytes are read from youtube.
-        '''
-
-        #! This will be called until download is complete, i.e we get an overall
-        #! self.progressBar.update(90)
-
-        fileSize = stream.filesize
-
-        #! How much of the file was downloaded this iteration scaled put of 90.
-        #! It's scaled to 90 because, the arbitrary division of each songs 100
-        #! iterations is (a) 90 for download (b) 5 for conversion & normalization
-        #! and (c) 5 for ID3 tag embedding
-        iterFraction = len(chunk) / fileSize * 90
-
-        self.progress += iterFraction
-        # if not int(self.progress) % 10:
-        self.send_update('Downloading...')
-
-    def notify_download_completion(self) -> None:
-        '''
-        updates progresbar to reflect a download being completed. 
-        '''
-        self.progress = 90
-        self.send_update('Download Complete, Converting....')
-    
-    def notify_conversion_completion(self) -> None:
-        '''
-        updates progresbar to reflect a audio conversion being completed
-        '''
-
-        self.progress = 95
-        self.send_update('Conversion Complete, Tagging...')
-
-    def notify_finished(self):
-        '''
-        updates progres to show the process is completed
-        '''
-        self.isComplete = True
-        self.progress = 100
-        self.send_update('Finished Tagging')
-        time.sleep(0.01) # Or else the messages will not be distinguishable from a timestamp sort
-        self.send_update('Done')
-
-    def send_update(self, message = ''):
-        '''
-        Called everytime the user should be notified.
-        This is where the communication information layout is defined.
-        queue -> { (processID): { 'time': (timestamp), 'name': (display name), 'progress': (int: 0-100), 'message': (str: message) } }
-        '''
-        if self.parent.lock:
-            self.parent.lock.acquire()
-        # self.parent.message_queue.put(line2) #, block=False
-        self.parent.put(self.ID, self.displayName, self.progress, message)
-        if self.parent.lock:
-            self.parent.lock.release()
-
-            
-    def close(self) -> None:
-        '''
-        clean up and exit
-        '''
-
-        self.send_update('Leaving Display Manager')
 
 
 
@@ -244,7 +72,7 @@ queue = Queue()
 ProgressRootProcess.register('MessageQueue', callable=lambda: queue)
 lock = Lock()
 ProgressRootProcess.register('Lock', callable=lambda: lock)
-ProgressRootProcess.register('ProcessDisplayManager', BasicDisplayMessenger)
+ProgressRootProcess.register('ParentMessageTracker', ParentMessageTracker)
 
 #! You can now run the following code to work with both DisplayManagers and
 #! DownloadTrackers:
@@ -286,7 +114,7 @@ class DownloadManager():
         self.messageQueue = progressRoot.MessageQueue()
         # self.lock = progressRoot.Lock()
         
-        self.processDisplayManager = progressRoot.ProcessDisplayManager(self.messageQueue)
+        self.parentMessageTracker = progressRoot.ParentMessageTracker(self.messageQueue)
         self.downloadTracker = progressRoot.DownloadTracker()
 
         self.progressRoot = progressRoot
@@ -308,15 +136,15 @@ class DownloadManager():
 
         downloads the given song
         '''
-        self.processDisplayManager.set_song_count_to(1)
+        self.parentMessageTracker.set_song_count_to(1)
         self.downloadTracker.clear()
         self.downloadTracker.load_song_list([songObj])
 
-        # download_song(songObj, self.processDisplayManager.newMessageTracker(songObj.get_song_name()), self.downloadTracker)
+        # download_song(songObj, self.parentMessageTracker.newMessageTracker(songObj.get_song_name()), self.downloadTracker)
         results = self.workerPool.apply_async(
             func     = download_song,
             args     = (
-                songObj, self.processDisplayManager.newMessageTracker(songObj.get_display_name()), self.downloadTracker
+                songObj, self.parentMessageTracker.newMessageTracker(songObj.get_display_name()), self.downloadTracker
             )
         )
 
@@ -338,7 +166,7 @@ class DownloadManager():
 
         downloads the given songs in parallel
         '''
-        self.processDisplayManager.set_song_count_to(len(songObjList))
+        self.parentMessageTracker.set_song_count_to(len(songObjList))
         self.downloadTracker.clear()
         self.downloadTracker.load_song_list(songObjList)
 
@@ -346,7 +174,7 @@ class DownloadManager():
         results = self.workerPool.starmap_async(
             func     = download_song,
             iterable = (
-                (songObj, self.processDisplayManager.newMessageTracker(songObj.get_display_name()), self.downloadTracker)
+                (songObj, self.parentMessageTracker.newMessageTracker(songObj.get_display_name()), self.downloadTracker)
                     for songObj in songObjList
             )
         )
@@ -375,12 +203,12 @@ class DownloadManager():
         
         songObjList = self.downloadTracker.get_song_list()
 
-        self.processDisplayManager.set_song_count_to(len(songObjList))
+        self.parentMessageTracker.set_song_count_to(len(songObjList))
 
         results = self.workerPool.starmap_async(
             func     = download_song,
             iterable = (
-                (songObj, self.processDisplayManager.newMessageTracker(songObj.get_display_name()), self.downloadTracker)
+                (songObj, self.parentMessageTracker.newMessageTracker(songObj.get_display_name()), self.downloadTracker)
                     for songObj in songObjList
             )
         )
